Changed instructionfactory signatures to uints. Seems easier to deal with constants this way, since we're forced to use the full range of int32

+++ OperatingSystemSimulation/src/Memory/ProgramCache.cs
@@ -27,7 +27,7 @@ namespace OperatingSystemSimulation.src.Memory
 
         public Instruction GetInstruction(Int32 programCounter)
         {
-            Int32 instructionData = read(programCounter);
+            UInt32 instructionData = (UInt32)read(programCounter);
 
             return InstructionFactory.CreateInstruction(instructionData);     
         }
